Q4 refactor WIP (initial trainer class)
