+++ loss/__init__.py
@@ -1,0 +1,9 @@
+from .loss_tal import TALoss
+from .loss_tal_dual import DualTALoss
+from .loss_tal_triple import TripleTALoss
+
+__all__ = [
+    "TALoss",
+    "DualTALoss",
+    "TripleTALoss"
+]
