+++ utils/autobatch.py
@@ -3,8 +3,8 @@ from copy import deepcopy
 import numpy as np
 import torch
 
-from utils.general import LOGGER, colorstr
-from utils.torch_utils import profile
+from .general import LOGGER, colorstr
+from .torch_utils import profile
 
 
 def check_train_batch_size(model, imgsz=640, amp=True):
