+++ utils/metrics.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import torch
 
-from utils import TryExcept, threaded
+from .general import TryExcept, threaded
 
 
 def fitness(x):
