+++ utils/object_det_plot.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import seaborn as sns
 from matplotlib import pyplot as plt
-from . import threaded
+from .general import threaded
 from .object_det_eval import smooth
