+++ utils/plots.py
@@ -14,10 +14,9 @@ import seaborn as sn
 import torch
 from PIL import Image, ImageDraw, ImageFont
 
-from utils import TryExcept, threaded
-from utils.general import (CONFIG_DIR, FONT, LOGGER, check_font, check_requirements, clip_boxes, increment_path,
-                           is_ascii, xywh2xyxy, xyxy2xywh)
-from utils.metrics import fitness
+from .general import CONFIG_DIR, FONT, LOGGER, check_font, check_requirements, clip_boxes, increment_path, \
+    is_ascii, xywh2xyxy, xyxy2xywh, TryExcept, threaded
+from .metrics import fitness
 
 # Settings
 matplotlib.rc('font', **{'size': 11})
