+++ utils/torch_utils.py
@@ -14,8 +14,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.parallel import DistributedDataParallel as DDP
 
-from utils.general import LOGGER, check_version, colorstr, file_date, git_describe
-from utils.lion import Lion
+from .general import LOGGER, check_version, file_date, git_describe
 
 try:
     import thop  # for FLOPs computation
@@ -305,218 +304,3 @@ def copy_attr(a, b, include=(), exclude=()):
             continue
         else:
             setattr(a, k, v)
-
-
-def smart_optimizer(model, name='Adam', lr=0.001, momentum=0.9, decay=1e-5):
-    # YOLOv5 3-param group optimizer: 0) weights with decay, 1) weights no decay, 2) biases no decay
-    g = [], [], []  # optimizer parameter groups
-    bn = tuple(v for k, v in nn.__dict__.items() if 'Norm' in k)  # normalization layers, i.e. BatchNorm2d()
-    #for v in model.modules():
-    #    for p_name, p in v.named_parameters(recurse=0):
-    #        if p_name == 'bias':  # bias (no decay)
-    #            g[2].append(p)
-    #        elif p_name == 'weight' and isinstance(v, bn):  # weight (no decay)
-    #            g[1].append(p)
-    #        else:
-    #            g[0].append(p)  # weight (with decay)
-                
-    for v in model.modules():
-        if hasattr(v, 'bias') and isinstance(v.bias, nn.Parameter):  # bias (no decay)
-            g[2].append(v.bias)
-        if isinstance(v, bn):  # weight (no decay)
-            g[1].append(v.weight)
-        elif hasattr(v, 'weight') and isinstance(v.weight, nn.Parameter):  # weight (with decay)
-            g[0].append(v.weight)
-            
-        if hasattr(v, 'im'):
-            if hasattr(v.im, 'implicit'):           
-                g[1].append(v.im.implicit)
-            else:
-                for iv in v.im:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia'):
-            if hasattr(v.ia, 'implicit'):           
-                g[1].append(v.ia.implicit)
-            else:
-                for iv in v.ia:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im2'):
-            if hasattr(v.im2, 'implicit'):           
-                g[1].append(v.im2.implicit)
-            else:
-                for iv in v.im2:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia2'):
-            if hasattr(v.ia2, 'implicit'):           
-                g[1].append(v.ia2.implicit)
-            else:
-                for iv in v.ia2:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im3'):
-            if hasattr(v.im3, 'implicit'):           
-                g[1].append(v.im3.implicit)
-            else:
-                for iv in v.im3:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia3'):
-            if hasattr(v.ia3, 'implicit'):           
-                g[1].append(v.ia3.implicit)
-            else:
-                for iv in v.ia3:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im4'):
-            if hasattr(v.im4, 'implicit'):           
-                g[1].append(v.im4.implicit)
-            else:
-                for iv in v.im4:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia4'):
-            if hasattr(v.ia4, 'implicit'):           
-                g[1].append(v.ia4.implicit)
-            else:
-                for iv in v.ia4:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im5'):
-            if hasattr(v.im5, 'implicit'):           
-                g[1].append(v.im5.implicit)
-            else:
-                for iv in v.im5:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia5'):
-            if hasattr(v.ia5, 'implicit'):           
-                g[1].append(v.ia5.implicit)
-            else:
-                for iv in v.ia5:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im6'):
-            if hasattr(v.im6, 'implicit'):           
-                g[1].append(v.im6.implicit)
-            else:
-                for iv in v.im6:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia6'):
-            if hasattr(v.ia6, 'implicit'):           
-                g[1].append(v.ia6.implicit)
-            else:
-                for iv in v.ia6:
-                    g[1].append(iv.implicit)
-                    
-        if hasattr(v, 'im7'):
-            if hasattr(v.im7, 'implicit'):           
-                g[1].append(v.im7.implicit)
-            else:
-                for iv in v.im7:
-                    g[1].append(iv.implicit)
-        if hasattr(v, 'ia7'):
-            if hasattr(v.ia7, 'implicit'):           
-                g[1].append(v.ia7.implicit)
-            else:
-                for iv in v.ia7:
-                    g[1].append(iv.implicit)
-
-    if name == 'Adam':
-        optimizer = torch.optim.Adam(g[2], lr=lr, betas=(momentum, 0.999))  # adjust beta1 to momentum
-    elif name == 'AdamW':
-        optimizer = torch.optim.AdamW(g[2], lr=lr, betas=(momentum, 0.999), weight_decay=0.0, amsgrad=True)
-    elif name == 'RMSProp':
-        optimizer = torch.optim.RMSprop(g[2], lr=lr, momentum=momentum)
-    elif name == 'SGD':
-        optimizer = torch.optim.SGD(g[2], lr=lr, momentum=momentum, nesterov=True)
-    elif name == 'LION':
-        optimizer = Lion(g[2], lr=lr, betas=(momentum, 0.99), weight_decay=0.0)
-    else:
-        raise NotImplementedError(f'Optimizer {name} not implemented.')
-
-    optimizer.add_param_group({'params': g[0], 'weight_decay': decay})  # add g0 with weight_decay
-    optimizer.add_param_group({'params': g[1], 'weight_decay': 0.0})  # add g1 (BatchNorm2d weights)
-    LOGGER.info(f"{colorstr('optimizer:')} {type(optimizer).__name__}(lr={lr}) with parameter groups "
-                f"{len(g[1])} weight(decay=0.0), {len(g[0])} weight(decay={decay}), {len(g[2])} bias")
-    return optimizer
-
-
-def smart_hub_load(repo='ultralytics/yolov5', model='yolov5s', **kwargs):
-    # YOLOv5 torch.hub.load() wrapper with smart error/issue handling
-    if check_version(torch.__version__, '1.9.1'):
-        kwargs['skip_validation'] = True  # validation causes GitHub API rate limit errors
-    if check_version(torch.__version__, '1.12.0'):
-        kwargs['trust_repo'] = True  # argument required starting in torch 0.12
-    try:
-        return torch.hub.load(repo, model, **kwargs)
-    except Exception:
-        return torch.hub.load(repo, model, force_reload=True, **kwargs)
-
-
-def smart_resume(ckpt, optimizer, ema, epochs):
-    # Resume training from a partially trained checkpoint
-    best_fitness = 0.0
-    start_epoch = 0
-    if ckpt.get('epoch') and isinstance(ckpt.get('epoch'), int):
-        start_epoch = ckpt['epoch'] + 1
-    if optimizer and ckpt.get('optimizer') and ckpt.get('best_fitness'):
-        optimizer.load_state_dict(ckpt['optimizer'])  # optimizer
-        best_fitness = ckpt['best_fitness']
-    if ema and ckpt.get('ema') and ckpt.get('updates'):
-        ema.ema.load_state_dict(ckpt['ema'].float().state_dict())  # EMA
-        ema.updates = ckpt['updates']
-    if epochs < start_epoch:
-        assert start_epoch < epochs, f"Start epoch {start_epoch} from checkpoint cannot be higher than total epochs {epochs}."
-    return best_fitness, start_epoch
-
-
-class EarlyStopping:
-    # YOLOv5 simple early stopper
-    def __init__(self, patience=30):
-        self.best_fitness = 0.0  # i.e. mAP
-        self.best_epoch = 0
-        self.patience = patience or float('inf')  # epochs to wait after fitness stops improving to stop
-        self.possible_stop = False  # possible stop may occur next epoch
-
-    def __call__(self, epoch, fitness):
-        if fitness >= self.best_fitness:  # >= 0 to allow for early zero-fitness stage of training
-            self.best_epoch = epoch
-            self.best_fitness = fitness
-        delta = epoch - self.best_epoch  # epochs without improvement
-        self.possible_stop = delta >= (self.patience - 1)  # possible stop may occur next epoch
-        stop = delta >= self.patience  # stop training if patience exceeded
-        if stop:
-            LOGGER.info(f'Stopping training early as no improvement observed in last {self.patience} epochs. '
-                        f'Best results observed at epoch {self.best_epoch}, best model saved as best.pt.\n'
-                        f'To update EarlyStopping(patience={self.patience}) pass a new patience value, '
-                        f'i.e. `python train.py --patience 300` or use `--patience 0` to disable EarlyStopping.')
-        return stop
-
-
-class ModelEMA:
-    """ Updated Exponential Moving Average (EMA) from https://github.com/rwightman/pytorch-image-models
-    Keeps a moving average of everything in the model state_dict (parameters and buffers)
-    For EMA details see https://www.tensorflow.org/api_docs/python/tf/train/ExponentialMovingAverage
-    """
-
-    def __init__(self, model, decay=0.9999, tau=2000, updates=0):
-        # Create EMA
-        self.ema = deepcopy(de_parallel(model)).eval()  # FP32 EMA
-        self.updates = updates  # number of EMA updates
-        self.decay = lambda x: decay * (1 - math.exp(-x / tau))  # decay exponential ramp (to help early epochs)
-        for p in self.ema.parameters():
-            p.requires_grad_(False)
-
-    def update(self, model):
-        # Update EMA parameters
-        self.updates += 1
-        d = self.decay(self.updates)
-
-        msd = de_parallel(model).state_dict()  # model state_dict
-        for k, v in self.ema.state_dict().items():
-            if v.dtype.is_floating_point:  # true for FP16 and FP32
-                v *= d
-                v += (1 - d) * msd[k].detach()
-        # assert v.dtype == msd[k].dtype == torch.float32, f'{k}: EMA {v.dtype} and model {msd[k].dtype} must be FP32'
-
-    def update_attr(self, model, include=(), exclude=('process_group', 'reducer')):
-        # Update EMA attributes
-        copy_attr(self.ema, model, include, exclude)
